refactor(media): the two files the HDR variant pushed past 300 lines, split
`kroma-domain/src/media.rs` reached 400 lines and `kroma-db/src/hydrate.rs` 302,
both over the hard limit, so both split at a seam rather than at a line count.

`media.rs` was holding three things: which filename tokens name a cut of a title,
what one track inside a file is, and the catalog entities themselves. It becomes
`media/{edition,streams,mod}.rs`. `mod.rs` re-exports the two, so every
`kroma_domain::VideoStream` and `detect_edition` path outside the crate resolves
unchanged.

`hydrate.rs` was loading an item's files and, separately, deciding which of them
represents it and mirroring that one onto the item's top-level fields. The second
half, and its tests, move to `hydrate/representative.rs`.

rustfmt over the two probe files this branch wrote, which it wanted reflowed.
Nothing else is reformatted.
