--- server/crates/kroma-domain/src/media.rs
@@ -1,400 +1,0 @@
-//! Media catalog types: kinds, stream descriptions, files, items and shows.
-//!
-//! The JSON shape here is a public contract web/TV clients depend on it, so
-//! field names and casing must not drift.
-
-use serde::{Deserialize, Serialize};
-
-use crate::metadata::{CastMember, Metadata};
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "lowercase")]
-pub enum Kind {
-    Movie,
-    Episode,
-    Video,
-}
-
-// (needle, label) cut/edition labels first, then source/quality.
-const EDITION_TABLE: &[(&str, &str)] = &[
-    ("director's cut", "Director's Cut"),
-    ("directors cut", "Director's Cut"),
-    ("director.cut", "Director's Cut"),
-    ("extended", "Extended"),
-    ("uncut", "Uncut"),
-    ("unrated", "Unrated"),
-    ("theatrical", "Theatrical"),
-    ("remastered", "Remastered"),
-    ("imax", "IMAX"),
-    ("remux", "Remux"),
-    ("2160p", "4K"),
-    ("4k", "4K"),
-    ("uhd", "4K"),
-    ("1080p", "1080p"),
-    ("720p", "720p"),
-    ("480p", "480p"),
-];
-
-/// The labels that name a distinct CUT of a title rather than a quality tier.
-/// Two files sharing a cut are the same content: one can replace the other.
-/// Two files differing in cut are different content and never replace one
-/// another, however their quality compares.
-pub const EDITION_CUTS: &[&str] = &[
-    "Director's Cut",
-    "Extended",
-    "Uncut",
-    "Unrated",
-    "Theatrical",
-    "Remastered",
-    "IMAX",
-];
-
-/// The edition label a file name carries, e.g. `Extended` or `4K`.
-pub fn detect_edition(file_name: &str) -> Option<String> {
-    let lower = file_name.to_ascii_lowercase();
-    EDITION_TABLE
-        .iter()
-        .find(|(needle, _)| lower.contains(needle))
-        .map(|(_, label)| label.to_string())
-}
-
-/// The cut an edition names, or `None` when it only names a quality tier
-/// (`4K`, `1080p`, `Remux`, ...) or nothing at all.
-pub fn edition_cut(edition: Option<&str>) -> Option<&'static str> {
-    let edition = edition?;
-    EDITION_CUTS
-        .iter()
-        .copied()
-        .find(|cut| cut.eq_ignore_ascii_case(edition))
-}
-
-/// The HDR system a video stream carries. Absent is SDR. Each has its own client
-/// support, so a set that renders HDR10 and a set that also renders Dolby Vision
-/// are not the same set.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum HdrFormat {
-    Hdr10,
-    Hdr10Plus,
-    DolbyVision,
-    Hlg,
-}
-
-impl HdrFormat {
-    pub fn as_str(self) -> &'static str {
-        match self {
-            HdrFormat::Hdr10 => "hdr10",
-            HdrFormat::Hdr10Plus => "hdr10Plus",
-            HdrFormat::DolbyVision => "dolbyVision",
-            HdrFormat::Hlg => "hlg",
-        }
-    }
-    pub fn parse(s: &str) -> Option<HdrFormat> {
-        match s {
-            "hdr10" => Some(HdrFormat::Hdr10),
-            "hdr10Plus" => Some(HdrFormat::Hdr10Plus),
-            "dolbyVision" => Some(HdrFormat::DolbyVision),
-            "hlg" => Some(HdrFormat::Hlg),
-            _ => None,
-        }
-    }
-}
-
-/// Colour signalling as the container states it, in ffprobe's own spelling
-/// (`bt2020`, `smpte2084`, `bt709`, ...) rather than a normalised one, because a
-/// client matches the exact value and a rename would lose what was read.
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct ColorInfo {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub primaries: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub transfer: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub matrix: Option<String>,
-}
-
-impl ColorInfo {
-    pub fn is_empty(&self) -> bool {
-        self.primaries.is_none() && self.transfer.is_none() && self.matrix.is_none()
-    }
-}
-
-/// Video stream description (best-effort; fields may be null when unknown).
-/// `hdr` is what a client that predates `hdr_format` reads: true for any variant,
-/// and still true where the variant itself was never recorded, which is how a row
-/// probed before the column existed keeps reading as HDR. So `hdr` without
-/// `hdr_format` means "HDR, variant unknown", never SDR.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct VideoStream {
-    pub codec: String,
-    pub width: Option<u32>,
-    pub height: Option<u32>,
-    pub hdr: bool,
-    #[serde(rename = "bitDepth")]
-    pub bit_depth: Option<u32>,
-    #[serde(rename = "hdrFormat", default, skip_serializing_if = "Option::is_none")]
-    pub hdr_format: Option<HdrFormat>,
-    #[serde(
-        rename = "dolbyVisionProfile",
-        default,
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub dolby_vision_profile: Option<u32>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub color: Option<ColorInfo>,
-}
-
-/// One audio stream/track. An item can carry several (e.g. EN + FR, or a
-/// director's commentary); `index` is the audio-relative position (0-based
-/// among audio streams only), matching ffmpeg's `-map 0:a:<index>` selector
-/// used when remuxing a chosen track.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AudioStream {
-    #[serde(default)]
-    pub index: u32,
-    pub codec: String,
-    pub channels: Option<u32>,
-    pub language: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub title: Option<String>,
-    #[serde(default)]
-    pub default: bool,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SubtitleTrack {
-    pub language: Option<String>,
-    pub codec: String,
-}
-
-/// Outcome of the EBU R128 loudness analysis of an audio track.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub enum AudioVerdict {
-    Ok,
-    HighDynamics,
-    QuietDialog,
-}
-
-/// EBU R128 loudness measurement of an item's default audio track, produced by
-/// the `pipeline.loudness` stage.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct AudioAnalysis {
-    #[serde(rename = "lufsI")]
-    pub lufs_i: f64,
-    // Loudness range (LU); > ~15 is the classic "quiet dialogue, loud
-    // explosions" mix.
-    pub lra: f64,
-    #[serde(rename = "truePeak")]
-    pub true_peak: f64,
-    // Centre-channel loudness (LUFS), measured for 5.1+ tracks only.
-    #[serde(
-        rename = "dialogLufs",
-        default,
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub dialog_lufs: Option<f64>,
-    pub verdict: AudioVerdict,
-}
-
-/// One physical file backing a logical [`MediaItem`]. A single item can have
-/// several of these (Director's Cut + Theatrical, 1080p + 4K, …); they all share
-/// the same logical item id but each maps to a distinct file on disk.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MediaFile {
-    // `short_hash(abs_path)`, stable per physical file.
-    pub id: String,
-    #[serde(rename = "relPath")]
-    pub rel_path: Option<String>,
-    pub container: String,
-    #[serde(rename = "durationMs")]
-    pub duration_ms: Option<u64>,
-    pub video: Option<VideoStream>,
-    pub audio: Option<AudioStream>,
-    #[serde(rename = "audioTracks", default)]
-    pub audio_tracks: Vec<AudioStream>,
-    pub subtitles: Vec<SubtitleTrack>,
-    pub size: Option<u64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub edition: Option<String>,
-    // `false` until ffprobe has run (phase 2); the stream fields above are
-    // null until then.
-    pub probed: bool,
-    // ffprobe's own reason the container would not open.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub unreadable: Option<String>,
-    #[serde(skip)]
-    pub abs_path: Option<String>,
-}
-
-/// A single playable media item. `rel_path` is relative to the owning media
-/// directory; demo/seed items have `rel_path == None` and cannot be streamed.
-///
-/// An item can be backed by multiple physical [`MediaFile`]s; the top-level
-/// `video`/`audio`/`duration_ms`/`container`/`subtitles`/`abs_path` fields
-/// mirror the highest-resolution probed file, for clients that read
-/// `item.video.codec` directly.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct MediaItem {
-    pub id: String,
-    pub title: String,
-    pub kind: Kind,
-    pub year: Option<u32>,
-    #[serde(rename = "durationMs")]
-    pub duration_ms: Option<u64>,
-    pub container: String,
-    pub video: Option<VideoStream>,
-    pub audio: Option<AudioStream>,
-    #[serde(rename = "audioTracks", default)]
-    pub audio_tracks: Vec<AudioStream>,
-    pub subtitles: Vec<SubtitleTrack>,
-    pub library: String,
-    #[serde(rename = "showId")]
-    pub show_id: Option<String>,
-    #[serde(rename = "showTitle")]
-    pub show_title: Option<String>,
-    pub season: Option<u32>,
-    pub episode: Option<u32>,
-    #[serde(rename = "episodeEnd")]
-    pub episode_end: Option<u32>,
-    #[serde(rename = "episodeTitle")]
-    pub episode_title: Option<String>,
-    #[serde(rename = "relPath")]
-    pub rel_path: Option<String>,
-    #[serde(rename = "addedAt")]
-    pub added_at: String,
-    // `None` until the background enrichment pass resolves it. Movies only;
-    // episodes inherit their show's metadata.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub metadata: Option<Metadata>,
-    // Mirrors the representative file's path so `/stream` keeps working.
-    #[serde(skip)]
-    pub abs_path: Option<String>,
-    #[serde(default)]
-    pub files: Vec<MediaFile>,
-    // Id of the representative file `/stream` serves and whose stream info
-    // populates the top-level fields above. `None` until a file exists.
-    #[serde(
-        rename = "defaultFileId",
-        default,
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub default_file_id: Option<String>,
-    // Episodes only. Empty until resolved from chapters or the
-    // audio-fingerprint job.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub markers: Vec<Marker>,
-    // `None` until the `pipeline.loudness` stage has measured it.
-    #[serde(
-        rename = "audioAnalysis",
-        default,
-        skip_serializing_if = "Option::is_none"
-    )]
-    pub audio_analysis: Option<AudioAnalysis>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[serde(rename_all = "lowercase")]
-pub enum MarkerKind {
-    Intro,
-    Credits,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Marker {
-    pub kind: MarkerKind,
-    #[serde(rename = "startMs")]
-    pub start_ms: u64,
-    #[serde(rename = "endMs")]
-    pub end_ms: u64,
-}
-
-/// A TV show aggregate (not a file). Built by grouping episodes during a scan.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Show {
-    pub id: String,
-    pub title: String,
-    pub year: Option<u32>,
-    pub library: String,
-    #[serde(rename = "seasonCount")]
-    pub season_count: u32,
-    #[serde(rename = "episodeCount")]
-    pub episode_count: u32,
-    // From a representative episode, for quality badges.
-    pub video: Option<VideoStream>,
-    #[serde(rename = "addedAt")]
-    pub added_at: String,
-    // `None` until the background enrichment pass resolves it.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub metadata: Option<Metadata>,
-    // Series-completion percent (0-100) when the request is authenticated;
-    // `None` for anonymous requests or shows with no progress.
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub progress: Option<u8>,
-}
-
-/// Sorted by episode number.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Season {
-    pub number: u32,
-    pub episodes: Vec<MediaItem>,
-    // Empty until enriched, or when the provider returned none.
-    #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub cast: Vec<CastMember>,
-}
-
-/// `GET /api/shows/:id` payload: a show plus its seasons.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct ShowDetail {
-    pub show: Show,
-    pub seasons: Vec<Season>,
-}
-
-/// One cover of the anonymous sign-in splash (`GET /api/splash`): enough to
-/// paint a backdrop and caption it, and deliberately nothing more.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct SplashEntry {
-    /// `movie` or `show`.
-    pub kind: String,
-    pub title: String,
-    pub year: Option<u32>,
-    #[serde(rename = "backdropUrl")]
-    pub backdrop_url: String,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub rating: Option<f64>,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn every_hdr_variant_stores_under_the_spelling_it_travels_as() {
-        for variant in [
-            HdrFormat::Hdr10,
-            HdrFormat::Hdr10Plus,
-            HdrFormat::DolbyVision,
-            HdrFormat::Hlg,
-        ] {
-            assert_eq!(
-                serde_json::to_string(&variant).unwrap(),
-                format!("\"{}\"", variant.as_str())
-            );
-            assert_eq!(HdrFormat::parse(variant.as_str()), Some(variant));
-        }
-
-        assert_eq!(HdrFormat::parse("hdr"), None);
-    }
-
-    #[test]
-    fn a_stream_with_nothing_to_say_about_colour_is_empty() {
-        let said_nothing = ColorInfo::default();
-        let said_something = ColorInfo {
-            transfer: Some("smpte2084".into()),
-            ..ColorInfo::default()
-        };
-
-        assert!(said_nothing.is_empty());
-        assert!(!said_something.is_empty());
-    }
-}
